==> Notebook1-PreProcess+FE.ipynb <==
#  Problem statement : 


- Build classifier that predicts the conversion-likelihood of a user


# Content

I have divided this assignment into five notebooks for clarity purpose. One notebook with so many lines of codes could be overwhelming for viewer.
In this notebook, we shall go through data basics and some data preprocessing techniques. 

- Loading csv files, merging useful information from iata dataset to main dataste
- Calculating geographic distance between origins and destinations
- Extracting new features out of original dataset
- Dealing with time series and time related variables & handing other data problems such as missing values
- Finally, saving data for next steps

# 1)-Importing key modules

# To avoid warnings Message 
import warnings
warnings.filterwarnings('ignore')
# For processing
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import scipy
import datetime as dt
from datetime import date
import datetime as dt
# For visualization
import seaborn as sns
plt.rcParams["figure.figsize"] = (16, 10)
plt.rcParams["xtick.labelsize"] = 10
plt.figure(figsize=(16,10)) # this creates a figure 16 inch wide, 10 inch high
from pprint import pprint
%matplotlib inline

# 2)-Loading data

df_event = pd.read_csv('events.csv')
df_event.shape

df_event.columns

df_event.head()

df_event.event_type.unique()

So we have a binary classification problem with our goal to predict the conversion-likelihood of a user

df_geo = pd.read_csv('iata_1_1.csv')

df_geo.shape

df_geo.head()

df_geo[df_geo.iata_code == "PAR"]

df_geo[df_geo.iata_code == "NYC"]

# 3)-Combining dataset

### a)- Working with orgin 

df_origin=pd.DataFrame(columns=["origin"])
df_origin['origin']=df_event['origin']
df_origin['origin_lat']=np.nan
df_origin.head(2)

for k in range(len(df_origin)):
    df_origin['origin_lat'][k] = (df_geo['lat'][df_geo['iata_code']==df_origin['origin'][k]][df_geo['lat'][df_geo['iata_code']==df_origin['origin'][k]].index[0]])

df_origin.head(2)

df_origin['origin_lon']=np.nan

for k in range(len(df_origin)):
    df_origin['origin_lon'][k] = (df_geo['lon'][df_geo['iata_code']==df_origin['origin'][k]][df_geo['lon'][df_geo['iata_code']==df_origin['origin'][k]].index[0]])

df_origin.head(2)

df_origin.shape

### b)-working with destination 

df_destination=pd.DataFrame(columns=["destination"])

df_destination['destination']=df_event['destination']
df_destination.head(2)

df_destination['destination_lat']=np.nan
df_destination['destination_lon']=np.nan

for k in range(len(df_destination)):
    df_destination['destination_lat'][k] = (df_geo['lat'][df_geo['iata_code']==df_destination['destination'][k]][df_geo['lat'][df_geo['iata_code']==df_destination['destination'][k]].index[0]])
    df_destination['destination_lon'][k] = (df_geo['lon'][df_geo['iata_code']==df_destination['destination'][k]][df_geo['lon'][df_geo['iata_code']==df_destination['destination'][k]].index[0]])

df_destination.head(2)

**Adding to main dataset**

df_event['origin_lat'] = df_origin['origin_lat']
df_event['origin_lon'] = df_origin['origin_lon']

df_event['destination_lat']= df_destination['destination_lat']
df_event['destination_lon']=df_destination['destination_lon']

df_event.head(2)

# 4)- Calculating distance

For geographic distance between origins and destinations, we choose Haversine distance between two points 

Reference for this : 

- https://www.movable-type.co.uk/scripts/latlong.html
- https://kanoki.org/2019/12/27/how-to-calculate-distance-in-python-and-pandas-using-scipy-spatial-and-distance-functions/

def haversine_vectorize(lon1, lat1, lon2, lat2):
 
    lon1, lat1, lon2, lat2 = map(np.radians, [lon1, lat1, lon2, lat2])
 
    newlon = lon2 - lon1
    newlat = lat2 - lat1
 
    haver_formula = np.sin(newlat/2.0)**2 + np.cos(lat1) * np.cos(lat2) * np.sin(newlon/2.0)**2
 
    dist = 2 * np.arcsin(np.sqrt(haver_formula ))
    km = 6367 * dist #6367 for distance in KM for miles use 3958
    return km

haversine_vectorize(df_event['origin_lon'],df_event['origin_lat'],df_event['destination_lon'],
                   df_event['destination_lat'])

df_event['distance'] = haversine_vectorize(df_event['origin_lon'],df_event['origin_lat'],df_event['destination_lon'],
                   df_event['destination_lat'])

df_event.head(2)

**Distance will be one of our key extracted feature for analysis**

# 5)- Creating new feature

num_family

df_event["num_family"]=df_event['num_adults']+df_event['num_children']

df_event.head(2)

drop extra columns for clarity. Also if we keep then they will cause multi-collinearity. W´shall us´extracted featurs for our analysis

df_event.drop(['num_adults', 'num_children','origin_lat','origin_lon','destination_lat','destination_lon'], axis=1, inplace=True)

df_event.head(2)

# 6)-Working with Time Series

Let's suppose. we want to see which year we have least/maximum online travel activities i.e search or booking. 

# Checking all variables with their data-type
def _tbl_dtype(dataset):
    sum_dtype = pd.DataFrame(dataset.dtypes).sort_values(0).rename(columns = {0:'Data Type'})
    return sum_dtype

table_dtype = _tbl_dtype(df_event)
table_dtype

### 6.1)-convert object to datetime

df_event["ts_datetime"]=pd.to_datetime(df_event.ts)
df_event["date_from_datetime"]=pd.to_datetime(df_event.date_from)
df_event["date_to_datetime"]=pd.to_datetime(df_event.date_to)

table_dtype = _tbl_dtype(df_event)
table_dtype

df_event.head(2)

**drop extra columns**

df_event.drop(['ts', 'date_to','date_from'], axis=1, inplace=True)

df_event.head(2)

### 6.2)- Handling missing values

df_event.isnull().sum()

We need to solve this missing value issue before anything

# to check what ate exact data samples that contain missing values
df_event[df_event["date_from_datetime"].isnull()]

these are our points where we get missing values.

**Solution**

Some like to remove missing values. I am not one of them. I like to keep my data and this data has already get very close to my heart. So, I ll keep data.

For keeping data, I can use impute or fill method. Impute is mostly used if missing data is in numeric variable column then we use "mean" or "median" to fill value.

I will use ffill i.e forward fill to deal with data. There are many other ways but, it is simple and we do not have many missing values in our data. We have only 22 considering out of 47007......is a small number. 


df_event=df_event.fillna(method="ffill")

df_event.isnull().sum()

### 6.3)-Adding extra Time Feature

- **journey_length (in Days)**

df_event['date_from_year']=df_event.date_from_datetime.dt.year
df_event['date_from_month']=df_event.date_from_datetime.dt.month
df_event['date_from_day']=df_event.date_from_datetime.dt.day
df_event['date_from_dayofweek']=df_event.date_from_datetime.dt.dayofweek

df_event.head()

df_event.date_from_year.unique()

df_event.date_from_month.unique()

df_event['date_to_year']=df_event.date_to_datetime.dt.year
df_event['date_to_month']=df_event.date_to_datetime.dt.month
df_event['date_to_day']=df_event.date_to_datetime.dt.day
df_event['date_to_dayofweek']=df_event.date_to_datetime.dt.dayofweek

df_event.date_to_year.unique()

df_event.head()

df_event.date_to_day.unique()

df_event['len_jour'] = df_event['date_to_datetime'].sub(df_event['date_from_datetime'], axis=0)

df_event[["date_from_datetime","date_to_datetime","len_jour"]].head()

Now we get how long is expcted journey of our customer. But , it is in days. For analysis, we need numeric form 

df_event['len_jour'] = df_event['len_jour'] / np.timedelta64(1, 'D')

df_event[["date_from_datetime","date_to_datetime","len_jour"]].head()

df_event[["date_from_datetime","date_to_datetime","len_jour"]].tail()

This information is interesting as it may describe our customer travelling habit. We may use this also as key extracted variable for our analysis

df_event.head()

**drop extra columns**

df_event.drop(['date_from_datetime', 'date_to_datetime',"date_from_year","date_from_month","date_from_day","date_from_dayofweek"], axis=1, inplace=True)
df_event.drop(["date_to_year","date_to_month","date_to_day","date_to_dayofweek"], axis=1, inplace=True)

df_event.head(2)

**For our analysis, we shall use three key features**

- num_family (in number)
- len_jour (in days)
- distance (in km)

target variable will be same ofcourse.

# 7)- Saving data

df_event.to_csv('travel_data.csv',index=False)

**END OF NOTEBOOK1**
