==> Notebook3.Class_imb+Model.ipynb <==
# Prediction of customers' travel pattern

# Content

- Loading data.
- Model Building.
- Dealing with imbalnced class problem.
- Choosing most relevant features.
- Using train-test split and also K-fold approaches for model evaluation, perfomance and validation.

# 1)-Importing key modules

import warnings
warnings.filterwarnings('ignore')
# For processing
import pandas as pd
import numpy as np
import scipy
import datetime as dt
from datetime import datetime
import seaborn as sns
import matplotlib.pyplot as plt
from pprint import pprint
%matplotlib inline

# For modeling building and validation
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import cross_val_score

# for evaluation

from sklearn import metrics
from sklearn.metrics import accuracy_score
from sklearn.metrics import recall_score
from sklearn.metrics import roc_auc_score
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix
from sklearn.model_selection import KFold
from sklearn.metrics import f1_score

# 2)-Loading and processing data

Loading from notebook2

df = pd.read_csv('selected_feature.csv')
df.shape

df.head()

df_model=df.copy()

df_model.head()

### spliting data into dependent and independent features

target=df_model["event_type"]
features=df_model.drop(['event_type'], axis=1)

print(target.shape)
print(features.shape)

features.head(2)

### normalize data

to deal with outlier issue seen in last notebook

from sklearn.preprocessing import StandardScaler
X = StandardScaler().fit_transform(features)

# 3)-Model Building

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, target, test_size=0.3, random_state=0)

print(X_train.shape)
print(X_test.shape)

print(y_train.shape)
print(y_test.shape)

X_train

# Logistic Classifeir
logreg = LogisticRegression(C=1e5)
logreg.fit(X_train, y_train)
predictions_LR = logreg.predict(X_test)

predictions_LR[:5]

print(accuracy_score(y_test,predictions_LR))

print(roc_auc_score(y_test, predictions_LR))

print(classification_report(y_test,predictions_LR))

We can clearly see that there is problem of imbalanced class. For booking results , we get very poor precision score of only 0.03 and for search events, we have 0.96. We need to solve this problem as we want a model that performs more precisely on booking instances than search results.

As for accuracy, it is very good. But, this does not make much value for our case. Also it makes case that accuracy is not always a good evaluation matric for checking model performance.

### Confusion matrix

C = confusion_matrix(y_test, predictions_LR)

C

This model does predict only 1 correct booking instance whereas 524 booking instances are incorrectly predicted as search instances. This is really bad for our case. We need a model that could predict booking instance correctly. even though we have an accuracy of 95%, we fail to get our required results 

### Precicion matrix

P =(C/C.sum(axis=0))
P

### Recall Matrix

A =(((C.T)/(C.sum(axis=1))).T)
A

# 4)- Solution to imbalanced Class

### Solution1: Under-sampling method

booking = df_model[df_model['event_type']==1]

search = df_model[df_model['event_type']==0]

target=df_model["event_type"]
features=df_model.drop(['event_type'], axis=1)
X = StandardScaler().fit_transform(features)
X_train, X_test, y_train, y_test = train_test_split(X, target, test_size=0.3, random_state=0)

from imblearn.under_sampling import NearMiss
# Implementing Undersampling for Handling Imbalanced 
nm = NearMiss()
X_under,y_under=nm.fit_sample(X_train,y_train)

print(X_under.shape,y_under.shape)

# checking classes
print(y_train.value_counts())

print(y_under.value_counts())

print(y_test.value_counts())

y_test.shape

# Logistic Classifeir
logreg = LogisticRegression()
logreg.fit(X_under, y_under)
predictions_LR_under = logreg.predict(X_test)

print(accuracy_score(predictions_LR_under,y_test))

print(classification_report(predictions_LR_under,y_test))

print(roc_auc_score(predictions_LR_under,y_test))

print(recall_score(predictions_LR_under,y_test))

print(f1_score(predictions_LR_under,y_test))

We could better results that our 1st model. Let's see what are other solutions in bag.

### Solution 2: Over-Sampling

adding artificial values to our training data. Mind you, we wont over sample test data. It will be a mistake as we only make our training data to learn and test that learning on test data.

from imblearn.combine import SMOTETomek

# Implementing Oversampling for Handling Imbalanced 
smk = SMOTETomek(random_state=42)
X_over,y_over=smk.fit_sample(X_train,y_train)

X_over.shape,y_over.shape

X_test.shape, y_test.shape

logreg = LogisticRegression(C=1e5)
logreg.fit(X_over, y_over)
predictions_LR = logreg.predict(X_test)

print(accuracy_score(predictions_LR,y_test))

print(roc_auc_score(predictions_LR,y_test))

print(classification_report(predictions_LR,y_test))

We can see that keeping a balanced class will give us a balanced precision result.It is still not very impressive on recall or f1 though.

### Solution 3: Using class_weight option as "balanced" 


using model for class balance

from sklearn.utils import class_weight

class_weights = class_weight.compute_class_weight('balanced',
                                                 np.unique(y_train),
                                                 y_train)

logreg = LogisticRegression(class_weight="balanced")

logreg.fit(X_train, y_train)
predictions_LR = logreg.predict(X_test)

print(accuracy_score(predictions_LR,y_test))

print(roc_auc_score(predictions_LR,y_test))

print(classification_report(predictions_LR,y_test))

We have three solutions. Oversample whether SMOTE or model option class balance give almost same results. With model option class balance, it does not seem like a big issue. But, I am going to choose under-sampling approach

### Why did we choose under sampling?

There are three reasons

- 1)- Our task is not to create a model for predicting accurate results for search and booking event classes. Our task is to find conversion-likelihood. Idea is to create a model that predicts very good results for booking events. If that model is not good in prediction of search event types then we won't mind. 

- 2)- As per my first interaction, our organization has got enough data. So, if we lose some samples then it is not big deal. I will consider it as first prototype round.In intital stage, we get less data and later we scale up project. If we get more data then this code will perform well for scaled data. Main concern is to deal with imbalanced classes and still to get good results for booking event type instances. That's why we can optimal model with our Solution 1. I have given other solutions as to know that I have knowledge of solutions.

- 3-If we apply more intense ML or DL models then over-sampling models will take alot of computing. If that is valuebale then sure , I would not mind high computation cost. But, in this assignment we can get pretty good results with under-sampling as well. 

# 4)- Choosing features

3 features vs six featues

y=df_model["event_type"]
features=df_model[["distance","num_family","len_jour","ts_hour"]]
X = StandardScaler().fit_transform(features)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=0)
nm = NearMiss()
X_under,y_under=nm.fit_sample(X_train,y_train)

### 4.1)- Using Statsmodel for Logistic Regression

import statsmodels.api as sm
logistic= sm.Logit(y_train,sm.add_constant(X_train)).fit()
logistic.summary2()

- ts_hour i. x4 is not significant as compare to p-value. So we can remove it from equation.
- num_family is marginal significant so, we shall keep it for analysis.

### 4.2)- Using sklearn for Logistic Regression

logreg = LogisticRegression()
logreg.fit(X_under, y_under)
predictions_LR = logreg.predict(X_test)

print(accuracy_score(predictions_LR,y_test))

print(classification_report(predictions_LR,y_test))

It is lower than our six feature model. So, it shows that those categorical variables(origin and destination) improved performance. We cannot leave them aside.

### 4.3)- Applying with removal of insignificant variable


using statsmodel

y=df_model["event_type"]
features=df_model[["distance","num_family","len_jour"]]
X = StandardScaler().fit_transform(features)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=0)
nm = NearMiss()
X_under,y_under=nm.fit_sample(X_train,y_train)

logistic= sm.Logit(y_train,sm.add_constant(X_train)).fit()
logistic.summary2()

### 4.4)- Using sklearn

logreg = LogisticRegression()
logreg.fit(X_under, y_under)
predictions_LR = logreg.predict(X_test)

print(accuracy_score(predictions_LR,y_test))

print(classification_report(predictions_LR,y_test))

Results are not same but, somewhat similar. Removal for one variable has not made much impact. 

still we will keep this three feature approached dropped and continue with five feature model. I am taking this liberty NOT to confine my model to three features as was directed in assignment.This is the only time I am deviating from assignments' instruction.

# 5)- Update Dataset

with five features in total

df.head()

df.drop(['ts_hour'], axis=1, inplace=True)

df.to_csv('updated_feature.csv',index=False)

# 6)- K-Fold 

Using balanced class data using SMOTE

from sklearn import metrics
df_model = pd.read_csv('updated_feature.csv')
y=df_model["event_type"]
features=df_model.drop(['event_type'], axis=1)
X = StandardScaler().fit_transform(features)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=0)

kf = KFold(n_splits=5)

for fold, (train_index, test_index) in enumerate(kf.split(X), 1):
    X_train = X[train_index]
    y_train = target[train_index]  # Based on our code, you might need a ravel call here, but I would look into how you're generating y
    X_test = X[test_index]
    y_test = target[test_index] 
    nm = NearMiss()
    X_under,y_under=nm.fit_sample(X_train,y_train)
    model = LogisticRegression()  # Choose a model here
    model.fit(X_under, y_under )  
    y_pred = model.predict(X_test)
    print(f'For fold {fold}:')
    print(f'Accuracy: {model.score(X_test, y_test)}')
    print(f'f-score: {f1_score(y_pred, y_test)}')
    print(f'roc-auc: {roc_auc_score(y_pred, y_test)}')
    print(f'recall: {recall_score(y_pred, y_test)}')
    print(f'precision: {metrics.precision_score(y_pred, y_test)}')

### cross validation done wrong

- Thanks to https://www.youtube.com/watch?v=DQC_YE3I5ig tutorial, I have found what is right and what is wrong way to apply K-fold.

kf = KFold(n_splits=5, random_state=42) <br>
accuracy = [] <br>
precision = [] <br>
recall = [] <br>
f1 = [] <br>
auc = [] <br>
X, y = SMOTE().fit_sample(X_train, y_train) <br>
for train, test in kf.split(X, y): <br>
    pipeline = make_pipeline(classifier(random_state=42)) <br>
    model = pipeline.fit(X[train], y[train]) <br>
    prediction = model.predict(X[test]) <br>
    accuracy.append(pipeline.score(X[test], y[test])) <br>
    precision.append(precision_score(y[test], prediction)) <br>
    recall.append(recall_score(y[test], prediction)) <br>
    f1.append(f1_score(y[test], prediction)) <br>

print("done wrong mean of scores 5-fold:") <br>
print("accuracy: {}".format(np.mean(accuracy))) <br>
print("precision: {}".format(np.mean(precision))) <br>
print("recall: {}".format(np.mean(recall))) <br>
print("f1: {}".format(np.mean(f1))) <br>

**Finally,** we get precison score of 0.75.These results are also very close to our earlier six feature model. But, would Regression Classifier be the only model ? We shall try more ML and DL model in next notebook to see better performance. To summarize ;

- Our dataset will be with five features i.e 3 numeric and 2 categorical
- We shall use under-sample method as per reason given above
- We shall try and compare results of Reg. Classifier with other models
- Our matric of evaluation will be precision.

**END OF NOTEBOOK3**
